Add BinaryReader and BinaryWriter classes.
Binary I/O is a bit of a boilerplate thing, so why not centralize it
here instead of writing several different implementations.

BinaryReader is a bit of a cut-down version of XB2AssetTool's
StreamHelper class, with none of the routines to read the XB2AT specific
datatypes. I don't know how this will be dealt with when XB2AT upgrades
mcommmon versions, but it'll have to be dealt with somehow... (I'm
personally thinking a inline function taking a BinaryReader& should
work?)